+++ lib/include/bave/graphics/render_instance.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <bave/graphics/rgba.hpp>
+#include <bave/graphics/topology.hpp>
 #include <bave/graphics/transform.hpp>
 #include <span>
 
@@ -26,6 +27,7 @@ struct RenderPrimitive {
 	std::size_t ibo_offset{};
 	std::uint32_t vertices{};
 	std::uint32_t indices{};
+	Topology topology{Topology::eTriangleStrip};
 };
 
 inline auto RenderInstance::bake() const -> Baked { return Baked{.transform = transform.matrix(), .rgba = Rgba::to_linear(tint.to_vec4())}; }