+++ lib/include/bave/graphics/topology.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace bave {
+/// \brief Primitive topology.
+enum class Topology : int { eTriangleStrip, eLineStrip };
+} // namespace bave